Finalized Error page, Finish page, Survey Page, and edited create.sql to have snowboardsize be a decimal instead of int

+++ Lumi/init_data/create.sql
@@ -6,7 +6,7 @@ CREATE TABLE users (
     email VARCHAR(40),
     password VARCHAR(40),
     level INT,
-    snowboardsize INT,
+    snowboardsize DECIMAL,
     shoeSize DECIMAL,
     skier INT,
     snowboarder INT
